pyproject.toml
[project]
name = "rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-core>=0.7.5",
    "autogen-ext>=0.7.5",
    "flagembedding>=1.3.5",
    "loguru>=0.7.3",
    "openai>=2.14.0",
    "pyautogen>=0.10.0",
    "python-telegram-bot>=22.5",
    "qdrant-client>=1.16.2",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
]
